order-processor/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
psycopg2-binary==2.9.10
redis==6.2.0